Reworded intro
Per a comment from Kai, took out the "2" in openHAB 2 to set the
precedent that whenever we refer to openHAB we are referring to OH 2 and
we will use the 1.x when referring to OH 1.

Reworked some of the wording to tighten up the sentences and take out
some redundancies.

Edited the bullet points, per many writing style guides bullet points
are part of the previous sentence and not a sentence unto themselves so
I reworded accordingly.