Reworked to talk about openHAB
Simplified the wording and tightened up some of the sentences to take
out redundancies .

Reworked bullet points to no longer be their own sentences per many
English writing style guides.

Tried to take some of the "developer" focus off of the text to make it
more suitable for a general audience. I did this by trying to make
definitions simpler and eliminating certain technical words or
descriptions in place of simpler and more straight forward ones.

Signed-off-by: Rich Koshak <rlkoshak@gmail.com> (github: rkoshak)