Adding example about.html files to the guidelines
In the Code Guidelines section, when mentioning the about.html file there could be two examples:
 - about.html file without 3rd party licence information
 - about.html file with 3rd party licence information
It will be easy for the contributers - they will spend less time figuring how the file should look and make fewer mistakes.

There could be a check when a bundle is commited, not only if it contains an about.html file, but also if the about.html file contains one of the dates in the example about.html files. This will help if a contributer has forgotten to change the date in the file.

The example files could be choosen from: https://github.com/openhab/openhab2-addons/pull/1263/files

Signed-off-by: Kiril Atanasov <kiril.atanasov@musala.com>