Is is correct to set the IP address of the Bridge?
Is it correct to place the IP address here? If I do so, Paper UI shows me two Bridges even though I uesed the correct IP address of the bridge in my environment. The one I added in the Things shows up as "Hue Bridge UNINITIALIZED" and the original discovered in the Paper UI is still called "Philips hue (192.168.178.18) ONLINE".