Binding docs track feature naming convention
To keep links to binding docs as they were and to track the feature name of bindings.  The bindings/oh1 and bindings/oh2 subdirectories are now gone.

This PR supercedes #296.

Should persistence and actions be moved up a level from their current “oh1” directory, since they don’t use the “1” appended to the feature name?  If yes, what happens when new features for persistence or actions are added later in OH2?

Signed-off-by: John Cocula <john@cocula.com>