Add note on 'System started' trigger
The 'System started' trigger causes actions to execute when the system is started (i.e. a reboot).  But the trigger also fires when alterations are made to any rules files.  This could lead to unanticipated system behavior, since users may expect that the only time the trigger would fire is when the system initializes for the first time.

Sign-off: Signed-off-by: Brad Gilmer brad@gilmer.tv (github: bgilmer77)