Clarify Item/real world interaction
This is a second test.  I am still learning about GitHub/GitHub Desktop interaction.  The purpose of the test is to check the behavior of GitHub when making multiple commits on a single branch.

I have modified the language to clarify the interaction between Items and the real world through bindings.  While the existing language was okay, I hope this submission provides additional clarity  around the fact that Item states can be set by actions in the real world