Binding Configuration: improve readability
Changes have been made in wording and arrangement to improve readability and comprehension.
Not sure whether all this 1.x information needs to remain.
Removed information on persistence.  Will look to see if it should be added to other PR on STATE #507.

This complete the nit pass through the items.md file