[Items] Add section on Item state, fixes #507 (#515)
* Add Section on Item State; fixes #507

Propose to add a section on Item State to the documentation on Items to address #507.
Adding an Item state diagram would be very useful.
Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Make additions visible in GitHub "View" mode

Strange - adding the TODO line caused the text below to not be visible on GitHub.  Removing it.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Correct text formatting so that each sentence is on one line

The Item State section has been reformatted so that each sentence is on its own line.
This seems to be the norm for line formatting.
Inserted a TODO regarding generating a state diagram.

* Remove TODO note

The TODO note about adding a state diagram has been removed - it seems to cause the entire new Item State section to disappear - at least in GitHub 'View' mode.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Correct links to other relevant articles

Links were incorrect.  This commit resolves the problem.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Incorporate "Restoring State" info at end of items.md into this section

I removed the text at the end of the items.md file related to restoring state.
Incorporated them into this new section.

* Address comments from @ThomDietrich (first round)

All comments received from @ThomDietrich's first round have been addressed.
Ran linter and reformatted list items.

* Remove section on restoring states

The section on restoring states has been removed - it repeats information which is available elsewhere in the documentation.

* Address comments from Thom dated 8-OCT

-  removed backticks that were used incorrectly.
-  added links to Persistence and system-based-trigger articles
-  removed trailing periods from list items
-

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)