[Items] Improve comprehension (#512)
* Improve introductory paragraph

This is a test.
DO NOT approve this PR.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Clarify Item/real world interaction

This is a second test.  I am still learning about GitHub/GitHub Desktop interaction.  The purpose of the test is to check the behavior of GitHub when making multiple commits on a single branch.

I have modified the language to clarify the interaction between Items and the real world through bindings.  While the existing language was okay, I hope this submission provides additional clarity  around the fact that Item states can be set by actions in the real world

* Improve readability of Introduction section

Changed wording in the Introduction section to improve readability and comprehension.

* Correct spelling error and clarify Item/real world interaction

Edited introduction to correct spelling error and clarify Item/real-world interaction.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github: bgilmer77)

* Item Definition and Syntax: improve readability

Some wording in the Item Definition and Syntax section has been changed to improve readability.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github:bgilmer77)

* Type: improve readability

Some wording in the Type section has been changed to improve readability

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github:bgilmer77)

* Name: improve readability

Some wording in the Name section has been changed, and the section has been reorganized to improve comprehension.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github:bgilmer77)

* Name: Tweak group naming suggestions

Based upon comment from @ThomDetrich, tweak the first scheme for group names.

Signed-off-by: Brad Gilmer <brad@gilmer.tv> (github:bgilmer77)

* Label: improve readability

Changed text in the Label section to improve readability and comprehension

* State: improve readability

Made changes to improve readability and comprehension.

* Icons: improve readability

The Icons section has been modified to improve readability and comprehension.

* Group: improve readability

Changes have been made in wording and arrangement to improve readability and comprehension.

* Tags: improve readability

Changes have been made in wording and arrangement to improve readability and comprehension.

* Binding: improve readability

Changes have been made in wording and arrangement to improve readability and comprehension.

* Binding Configuration: improve readability

Changes have been made in wording and arrangement to improve readability and comprehension.
Not sure whether all this 1.x information needs to remain.
Removed information on persistence.  Will look to see if it should be added to other PR on STATE #507.

This complete the nit pass through the items.md file

* Correct linter formatting issues

This commit corrects formatting issues revealed by linter.

* Add some last minute changes

Signed-off-by: Thomas Dietrich <thomas.dietrich@tu-ilmenau.de> (github: ThomDietrich)

* Discuss distinction between command and status

-  End-users may assume that the change in state of a Switch Item reflects the actual state of the Thing.
-  Added section to make clear the difference between command and status
-  May seem obvious, but some implementers could get burned by not thinking this through.

* Add second batch of corrections

Signed-off-by: Thomas Dietrich <thomas.dietrich@tu-ilmenau.de> (github: ThomDietrich)

* Correct typo

Signed-off-by: Thomas Dietrich <thomas.dietrich@tu-ilmenau.de> (github: ThomDietrich)