Fix wording of UDP binding
There was a three-year old note in the documentation which nobody had clarified. I am not a developer, but just stumbled over the documentation simply not matching what 2.3 does so my wording fixes at least that. 

However: a developer  urgently needs to address whether all of this is still true ...