Clarify Contact
I spent hours trying to find the problem, as I was trying to send a command to a contact. I think it should be mentioned a contact item only accepts status, and the 'CommandType' should be renamed OpenClosed with a D as the status is in fact CLOSED and not CLOSE. Less confusing this way.