Updating to reflect changes since OH 2 pre release
Adopted the one sentence per line convention used everywhere else in the docs.

Rewrote the paragraph discussing the versions of OH 2 available (Release, Testing, SNAPSHOT). NOTE: the Installation Overview needs to be updated as well. 

Removed discussion of offline and online since I don't think that is correct anymore. There is no mention of different offline/online installs in the installation docs.

Tried to make it more clear that the text based instructions are the recommended.

Replaces 1.9.0 with 1.x.0 and replaced 2.0 with 2.x to make the instructions not specific to the version.

Removed the warnings to create the persistence folder when running in Docker. They are no longer necessary.

In necessary changes for Rules, I removed the "core" from the org.openhab imports as there are some users who might be using script imports as well.

Removed references to ESHD and changed them to reference VSCode instead. Solves #775 

Signed off by Richard Koshak rlkoshak@gmail.com