Add VuePress infrastructure for preview
This allows a preview VuePress resembling the final
website to be built in order to preview the main
documentation (in the openhab-docs repository, content
fetched from other repos is excluded).

Started to refactor the sidebar navigation items and
scripts processing pages for VuePress compatibility so
they can be maintained in the openhab-docs repository.

Signed-off-by: Yannick Schaus <github@schaus.net>