Added clarification to the AC governance
I added clarification that the clock doesn't start to tick on a vote until there is actually a call for a vote. 

Added a section to describe how the AC can generate an idea and elicit feedback and discussion from the community at large. This section also discusses how a member of the community can bring the attention of the AC to a topic that may need AC input.