Change description of System started description
Based on the discussion at https://community.openhab.org/t/systemstart-rule-trigger-if-i-change-any-item/74542 and after performing a few tests of my own including creating a brand new .items file with an Item not used by ANY Rule I have verified that all System started Rules trigger for any change to any .items file. The "related to that rule file" part of the description is misleading and incorrect and should be removed.