Update linux.md for Samba install
I have the openhabian (openhab v2.5.8-1) install on an RPi 4B and couldn't get network sharing working so went through this process to add Samba. At the end of it, my OPENHAB now shows up on network of my Mac. Anyway, when adding the openhab user (for Samba access) to the openhab folders, I couldn't find the /opt/openhab2/conf location. I think that the latter part of line 781 can be removed for this reason.