Update first_steps.md (#1412)
* Update first_steps.md

I don't think the "Log into the dashboard with an existing user" section is necessary in the First Steps, since the user is already logged in after creating an administrator and there are no other existing users. It might confuse people thinking that they're missing something, when they're already logged in. I'd suggest cutting this section (is "logging into the dashboard" covered anywhere else?) and skip right to "Initial Settings".

When I try to space out the images, some of them disappear in the preview. I'm not sure what I'm doing wrong.

* Update first_steps.md

Updated to put each sentence on a separate line.

I also rewrote the "Log into the dashboard" section to simplify it for new users, while keeping the note that additional users can only be created via the console.

Signed-off-by: Russell Wong, russ@scatterthought.com (github: scatterthought)