Removed the use of bridge in network
Bridge makes the installation of openhab on synology more complex, it limits the internal ip traffic to stay in the container. Bindings like Network Binding won't work easily.
Normally you would optimize for security, but here the choice was made to optimize for ease of use. The latter is also the case if openhab was installed directly on the synology like we did on oh2

Signed-off-by: André Scholten <36789802+TallDragon@users.noreply.github.com>