Update sitemaps.md (#1676)
* Update sitemaps.md

As a new user, I was frustrated to find this out far too late. I really did read, twice, the page about sitemaps. I didn't understand why I couldn't see the results as in the screenshots. Where is it? Am I crazy?

There is so much documentation, a new user is not going to read every page. They learn, read or hear that they would need a sitemap, so that's what they are going to read in the documentation. They are not going to read a [different page](https://www.openhab.org/docs/ui/) first, where the above warning is present.

I advise all documentation writers to keep, move or duplicate all important information surrounding a subject _on the page about that subject_.

* Use tip box

Signed-off-by: Jerome Luckenbach <github@luckenba.ch>

Co-authored-by: Jerome Luckenbach <github@luckenba.ch>