Changed apt installation instructions gpg import (#1845)
apt-key is deprecated and shouldn't be used nor should the trusted keystore in /etc/apt. Instead the gpg keys should be stored somewhere else and the list file that adds the openHAB repo should point at that one key. Thus, apt will only trust that key for that repo instead of trusting that key even if it's used by some other repo.

There were problems with the redirect from `gpg` which makes the command fail sometimes even when using sudo which is why I save it locally and then move and chmod it after.

Signed-off-by: Richard Koshak rlkoshak@gmail.com