Moved reverse-proxy configuration to it's own page
To give a better balance between the different options I moved the NGINX configuration to it's own page.
So that the security page is focussed on the different options, and not looks like a configuration page for NGINX as it's the preferred way to do this.
Also added the option to run your own openHAB cloud.

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>