configuration/items: use more specific item types
I think having all examples using as specific item types, as possible,
utilizing the default UoM, is a good practice.

The wording “If no state presentation and no square brackets are given…”
leaves unclear what happens, if state presentation is missing, but square
brackets are present.  This is refined here.